🚚 refactor file structure to adhere to Flutter package conventions
