+++ lib/src/models/cover_data.dart
