+++ lib/src/models/cover_style.dart
