+++ lib/src/models/trim_style.dart
