+++ lib/src/widgets/cover/cover_viewer.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:video_editor/domain/entities/cover_data.dart';
-import 'package:video_editor/domain/entities/transform_data.dart';
-import 'package:video_editor/domain/bloc/controller.dart';
-import 'package:video_editor/domain/helpers.dart';
-import 'package:video_editor/ui/crop/crop_mixin.dart';
+import 'package:video_editor/src/controller.dart';
+import 'package:video_editor/src/utils/helpers.dart';
+import 'package:video_editor/src/models/cover_data.dart';
+import 'package:video_editor/src/models/transform_data.dart';
+import 'package:video_editor/src/widgets/crop/crop_mixin.dart';
 
 class CoverViewer extends StatefulWidget {
   /// It is the viewer that show the selected cover
