+++ lib/src/widgets/crop/crop_grid_painter.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:video_editor/domain/entities/crop_style.dart';
-import 'package:video_editor/ui/crop/crop_grid.dart';
+import 'package:video_editor/src/models/crop_style.dart';
+import 'package:video_editor/src/widgets/crop/crop_grid.dart';
 
 class CropGridPainter extends CustomPainter {
   const CropGridPainter(
