+++ lib/src/widgets/transform.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:video_editor/domain/entities/transform_data.dart';
+import 'package:video_editor/src/models/transform_data.dart';
 
 class CropTransform extends StatelessWidget {
   const CropTransform({
