+++ lib/src/widgets/trim/trim_slider.dart
@@ -2,9 +2,9 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:video_editor/domain/bloc/controller.dart';
-import 'package:video_editor/ui/trim/thumbnail_slider.dart';
-import 'package:video_editor/ui/trim/trim_slider_painter.dart';
+import 'package:video_editor/src/controller.dart';
+import 'package:video_editor/src/widgets/trim/thumbnail_slider.dart';
+import 'package:video_editor/src/widgets/trim/trim_slider_painter.dart';
 
 enum _TrimBoundaries { left, right, inside, progress }
